01_ai_agents_first/03_litellm_openai_agent/03_Litellm_agent_sdk/pyproject.toml
[project]
name = "03-litellm-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents[litellm]>=0.0.19",
    "python-dotenv>=1.1.1",
]
